Add support for reflective lookups
To achieve this, I kept a track of unresolved reflective lookups (where
the record and relatedTo SObjectTypes are the same, the the relatedTo
doesn't yet have an Id). If any of the records being inserted have a
reflective reference to anoth
er record being inserted in the same commit, then I perform an update
immediatel
y after the insert to "Backfill" the lookups with the Ids of the newly
inserted
records.

The tradeoff is clear and calculable. SObjects with unresolved
relationships incur 2xDML rows. I like this better than trying to order
the inserting of SObjects and incur and undetermined number of DML
statements.