Fields that are polymorphic fail when the chosen sObjectType is not the type specified in the selector.
Add the ability to set the sObjectType when validating the field exists on the object.  Use the type set in the selector to determine which polymorphic type to use.

Example: Lead.Owner.UserRoleID
Owner is GROUP | USER
Group does not contain UserRoleID and fails when the getField call sets the sObjectField token to null.