#233 Fix for list out of bounds
When a query with relationship fields in the select is generated for a community user in a without sharing context on an object that is not accessible (CRUD), a list out of bounds error is returned.

The method Schema.getGlobalDescribe().get('MyObjectName__c').getDescribe().fields.getMap().get('MyRelationShipFieldToAnotherNonAccessibleObject__c').getDescribe().getReferenceTo() returns null, when executed as community user.